+++ projects/demo/src/app/app.config.ts
@@ -1,4 +1,4 @@
-import { ApplicationConfig } from '@angular/core';
+import { ApplicationConfig, provideExperimentalZonelessChangeDetection } from '@angular/core';
 import { provideRouter } from '@angular/router';
 
 import { provideClientHydration } from '@angular/platform-browser';
@@ -7,6 +7,8 @@ import { routes } from './app.routes';
 
 export const appConfig: ApplicationConfig = {
   providers: [
+    provideExperimentalZonelessChangeDetection(),
+    
     provideRouter(routes),
     provideClientHydration(),
     provideNgxCountAnimations({