+++ projects/ngx-count-animation/src/public-api.ts
@@ -1,5 +1,4 @@
 export * from './lib/ngx-count-animation.directive';
-export * from './lib/ngx-count-animation.module';
-export * from './lib/ngx-count-animation.provider';
 export * from './lib/ngx-count-animation.service';
+export * from './lib/provider/ngx-count-animation.provider';